Split training into dedicated Azure ML script
